add desktop switch and try suspend
Added a few new hotkeys for switching and controlling desktops on windows.
Started a new system to suspend typing script while plugin runs, but not working yet.

line 51-76: Pressing Numpad0 & Esc will now set the numlock state to what it is on the external numpad state. This is to compensate for a bug in that numpad's software and shouldn't be needed later.